Fixed a compiler warning that only shows up on linux

+++ src/core/ext/transport/cronet/transport/cronet_transport.c
@@ -218,7 +218,7 @@ static void on_write_completed(cronet_bidirectional_stream *stream,
 static void process_recv_message(stream_obj *s, const uint8_t *recv_data) {
   gpr_slice read_data_slice = gpr_slice_malloc((uint32_t)s->total_read_bytes);
   uint8_t *dst_p = GPR_SLICE_START_PTR(read_data_slice);
-  memcpy(dst_p, recv_data, s->total_read_bytes);
+  memcpy(dst_p, recv_data, (size_t)s->total_read_bytes);
   gpr_slice_buffer_add(&s->read_slice_buffer, read_data_slice);
   grpc_slice_buffer_stream_init(&s->sbs, &s->read_slice_buffer, 0);
   *s->recv_message = (grpc_byte_buffer *)&s->sbs;
